feat: lab4 done (initial)
Took 4 hours 17 minutes

+++ Banks/Models/TransactionLog.cs
@@ -4,7 +4,7 @@ namespace Banks.Models
 {
     public class TransactionLog
     {
-        public TransactionLog(IAccount accountFrom, IAccount accountTo, Bank bankFrom, Bank bankTo, int amount, TransactionTypes type)
+        public TransactionLog(IAccount accountFrom, IAccount accountTo, Bank bankFrom, Bank bankTo, double amount, TransactionTypes type)
         {
             AccountFrom = accountFrom;
             AccountTo = accountTo;
@@ -18,7 +18,7 @@ namespace Banks.Models
         public IAccount AccountTo { get; }
         public Bank BankFrom { get; }
         public Bank BankTo { get; }
-        public int Amount { get; }
+        public double Amount { get; }
         public TransactionTypes Type { get; }
     }
 }
